#1592 Fixing for the latest reviews
